Combine junit test
